Moved DetectorBackends to new files
